Remove unused steamcmd_output.txt file
The steamcmd_output.txt file has been deleted as this file is not used or referenced anywhere within the application. This file only contained logging information from a console client, which is not relevant or needed for the running of the current application. This helps to reduce project clutter and better maintain our software structure. It is auto generated by the script