Ensure proper self-update of SteamCMD
Added a routine that prepares a string of +ggs with a fixed count of 10. This allows SteamCMD to update itself and fetch the latest information using the app_info_print command. This circumvents any issues related to outdated SteamCMD binaries. Also, 'start_new_session' is now passed to 'subprocess.run' for a neat cleanup of the spawned process.